+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Changelog
 
 * Fixed updating dangerous function restrictions when restrict dangerous enabled ([issue](/../../issues/1742))
 * Display system components as disabled when restricting system components is not enabled ([issue](/../../issues/1744))
+* Display no usage data when restriction not available or disabled
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)